[N02] Usage of SafeMath in non-uint256 variables

+++ contracts/OptionsContract.sol
@@ -511,9 +511,9 @@ contract OptionsContract is Ownable, ERC20 {
         // calculate the maximum amount of collateral that can be liquidated
         uint256 maxCollateralLiquidatable = maxCollateralLiquidatable(vaultOwner);
         if(liquidationFactor.exponent > 0) {
-            maxCollateralLiquidatable = maxCollateralLiquidatable.mul(10 ** uint32(liquidationFactor.exponent));
+            maxCollateralLiquidatable = maxCollateralLiquidatable.mul(10 ** uint256(liquidationFactor.exponent));
         } else {
-            maxCollateralLiquidatable = maxCollateralLiquidatable.div(10 ** uint32(-1 * liquidationFactor.exponent));
+            maxCollateralLiquidatable = maxCollateralLiquidatable.div(10 ** uint256(-1 * liquidationFactor.exponent));
         }
 
         require(amtCollateralToPay <= maxCollateralLiquidatable,
@@ -592,14 +592,14 @@ contract OptionsContract is Ownable, ERC20 {
         uint256 rightSideVal = (collateralAmt.mul(ethToStrikePrice)).div(ethToCollateralPrice);
         int32 rightSideExp = collateralExp;
 
-        uint32 exp = 0;
+        uint256 exp = 0;
         bool isSafe = false;
 
         if(rightSideExp < leftSideExp) {
-            exp = uint32(leftSideExp - rightSideExp);
+            exp = uint256(leftSideExp - rightSideExp);
             isSafe = leftSideVal.mul(10**exp) <= rightSideVal;
         } else {
-            exp = uint32(rightSideExp - leftSideExp);
+            exp = uint256(rightSideExp - leftSideExp);
             isSafe = leftSideVal <= rightSideVal.mul(10 ** exp);
         }
